Updated appearance; v0.1.0 release

+++ src/main/java/dev/piste/vayna/api/henrik/HenrikAccount.java
@@ -30,7 +30,7 @@ public class HenrikAccount {
         return puuid;
     }
 
-    public long getLevel() {
+    public int getLevel() {
         return level;
     }
